@@ -4,16 +4,15 @@ version: 1.0.0+1
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.6.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
+  cupertino_icons: ^1.0.2
   flutter:
     sdk: flutter
-
-  cupertino_icons: ^0.1.2
-  http: ^0.12.0+1
   flutter_appauth:
     path: ../
+  http: ^0.13.0
 
 dev_dependencies:
   flutter_test:
