@@ -1,10 +1,10 @@
 name: flutter_appauth
 description: This plugin provides an abstraction around the Android and iOS AppAuth SDKs so it can be used to communicate with OAuth 2.0 and OpenID Connect providers
-version: 0.9.2+6
+version: 1.0.0-nullsafety.0
 homepage: https://github.com/MaikuB/flutter_appauth/tree/master/flutter_appauth
 
 environment:
-  sdk: '>=2.6.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
   flutter: '>=1.12.13+hotfix.5'
 
 dependencies:
@@ -15,6 +15,7 @@ dependencies:
     git:
       url: https://github.com/akdu12/flutter_appauth.git
       path: flutter_appauth_web
+
 
 flutter:
   plugin:
