@@ -2,18 +2,18 @@ import 'package:flutter_appauth_platform_interface/flutter_appauth_platform_inte
 
 class FlutterAppAuth {
   /// Convenience method for authorizing and then exchanges code
-  Future<AuthorizationTokenResponse> authorizeAndExchangeCode(
+  Future<AuthorizationTokenResponse?> authorizeAndExchangeCode(
       AuthorizationTokenRequest request) {
     return FlutterAppAuthPlatform.instance.authorizeAndExchangeCode(request);
   }
 
   /// Sends an authorization request
-  Future<AuthorizationResponse> authorize(AuthorizationRequest request) {
+  Future<AuthorizationResponse?> authorize(AuthorizationRequest request) {
     return FlutterAppAuthPlatform.instance.authorize(request);
   }
 
   /// For exchanging tokens
-  Future<TokenResponse> token(TokenRequest request) {
+  Future<TokenResponse?> token(TokenRequest request) {
     return FlutterAppAuthPlatform.instance.token(request);
   }
 }
