@@ -1,9 +1,8 @@
 class AuthorizationServiceConfiguration {
-  AuthorizationServiceConfiguration(
+  const AuthorizationServiceConfiguration(
     this.authorizationEndpoint,
     this.tokenEndpoint,
-  ) : assert(tokenEndpoint != null && authorizationEndpoint != null,
-            'Must specify both the authorization and token endpoints');
+  );
 
   final String authorizationEndpoint;
 
