@@ -6,9 +6,9 @@ import 'common_request_details.dart';
 import 'grant_types.dart';
 import 'token_request.dart';
 
-Map<String, Object> _convertCommonRequestDetailsToMap(
+Map<String, Object?> _convertCommonRequestDetailsToMap(
     CommonRequestDetails commonRequestDetails) {
-  return <String, Object>{
+  return <String, Object?>{
     'clientId': commonRequestDetails.clientId,
     'issuer': commonRequestDetails.issuer,
     'discoveryUrl': commonRequestDetails.discoveryUrl,
@@ -21,7 +21,7 @@ Map<String, Object> _convertCommonRequestDetailsToMap(
 }
 
 extension AuthorizationRequestParameters on AuthorizationRequest {
-  Map<String, Object> toMap() {
+  Map<String, Object?> toMap() {
     return _convertAuthorizationParametersToMap(this)
       ..addAll(_convertCommonRequestDetailsToMap(this));
   }
@@ -38,20 +38,20 @@ extension AuthorizationServiceConfigurationMapper
 }
 
 extension TokenRequestMapper on TokenRequest {
-  Map<String, Object> toMap() {
+  Map<String, Object?> toMap() {
     return _convertTokenRequestToMap(this);
   }
 }
 
 extension AuthorizationTokenRequestMapper on AuthorizationTokenRequest {
-  Map<String, Object> toMap() {
+  Map<String, Object?> toMap() {
     return _convertTokenRequestToMap(this)
       ..addAll(_convertAuthorizationParametersToMap(this));
   }
 }
 
-Map<String, Object> _convertTokenRequestToMap(TokenRequest tokenRequest) {
-  return <String, Object>{
+Map<String, Object?> _convertTokenRequestToMap(TokenRequest tokenRequest) {
+  return <String, Object?>{
     'clientSecret': tokenRequest.clientSecret,
     'refreshToken': tokenRequest.refreshToken,
     'authorizationCode': tokenRequest.authorizationCode,
@@ -60,7 +60,7 @@ Map<String, Object> _convertTokenRequestToMap(TokenRequest tokenRequest) {
   }..addAll(_convertCommonRequestDetailsToMap(tokenRequest));
 }
 
-String _inferGrantType(TokenRequest tokenRequest) {
+String? _inferGrantType(TokenRequest tokenRequest) {
   if (tokenRequest.grantType != null) {
     return tokenRequest.grantType;
   }
@@ -75,9 +75,9 @@ String _inferGrantType(TokenRequest tokenRequest) {
       null, 'grantType', 'Grant type not specified and cannot be inferred');
 }
 
-Map<String, Object> _convertAuthorizationParametersToMap(
+Map<String, Object?> _convertAuthorizationParametersToMap(
     AuthorizationParameters authorizationParameters) {
-  return <String, Object>{
+  return <String, Object?>{
     'loginHint': authorizationParameters.loginHint,
     'promptValues': authorizationParameters.promptValues,
     'preferEphemeralSession': authorizationParameters.preferEphemeralSession,
