@@ -27,7 +27,7 @@ void main() {
     expect(
       log,
       <Matcher>[
-        isMethodCall('authorize', arguments: <String, Object>{
+        isMethodCall('authorize', arguments: <String, Object?>{
           'clientId': 'someClientId',
           'issuer': null,
           'redirectUrl': 'someRedirectUrl',
@@ -51,7 +51,7 @@ void main() {
     expect(
       log,
       <Matcher>[
-        isMethodCall('authorizeAndExchangeCode', arguments: <String, Object>{
+        isMethodCall('authorizeAndExchangeCode', arguments: <String, Object?>{
           'clientId': 'someClientId',
           'issuer': null,
           'redirectUrl': 'someRedirectUrl',
@@ -87,7 +87,7 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('token', arguments: <String, Object>{
+          isMethodCall('token', arguments: <String, Object?>{
             'clientId': 'someClientId',
             'issuer': null,
             'redirectUrl': 'someRedirectUrl',
@@ -113,7 +113,7 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('token', arguments: <String, Object>{
+          isMethodCall('token', arguments: <String, Object?>{
             'clientId': 'someClientId',
             'issuer': null,
             'redirectUrl': 'someRedirectUrl',
@@ -138,7 +138,7 @@ void main() {
       expect(
         log,
         <Matcher>[
-          isMethodCall('token', arguments: <String, Object>{
+          isMethodCall('token', arguments: <String, Object?>{
             'clientId': 'someClientId',
             'issuer': null,
             'redirectUrl': 'someRedirectUrl',
